Refactor config.py into a clean facade over config_builders loading logic
Move all loading/resolution functions (load_hub_config, load_repo_config,
derive_repo_config, ensure_hub_config_at, etc.) from config.py into
config_builders.py. Break the bootstrap->config cycle by having bootstrap
import directly from config_layering/config_contract. The facade now has
zero E402 workaround imports and zero function-level import workarounds.
